app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
google-api-python-client==1.12.1
google-auth==2.26.2
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
python-multipart==0.0.6